Rework again how multiple handles are implemented.
Multiple problems with the previous solution became apparent. As a positive side effect it is no longer required to hardcode the number of handles as they are now dynamically added. A new node type is not necessary. Also, while adding new types is easy, these (seemingly) cannot have addition arguments, instead everything has to be carried in the data dictionary.